requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
numpy>=1.24.0
requests>=2.28.0
tqdm>=4.65.0
